requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
mysqlclient==2.0.3
numpy==1.21.0
pandas==1.3.0
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
